add group uuid from boards
